+++ cmd/get/clusters/provider/aws.go
@@ -1,7 +1,7 @@
 package provider
 
 import (
-	infrastructurev1alpha2 "github.com/giantswarm/apiextensions/v3/pkg/apis/infrastructure/v1alpha2"
+	infrastructurev1alpha3 "github.com/giantswarm/apiextensions/v3/pkg/apis/infrastructure/v1alpha3"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 
@@ -35,7 +35,7 @@ func GetAWSTable(clusterResource cluster.Resource) *metav1.Table {
 }
 
 func getAWSClusterRow(c cluster.Cluster) metav1.TableRow {
-	if c.V1Alpha2Cluster == nil || c.AWSCluster == nil {
+	if c.Cluster == nil || c.AWSCluster == nil {
 		return metav1.TableRow{}
 	}
 
@@ -54,7 +54,7 @@ func getAWSClusterRow(c cluster.Cluster) metav1.TableRow {
 	}
 }
 
-func getLatestAWSCondition(conditions []infrastructurev1alpha2.CommonClusterStatusCondition) string {
+func getLatestAWSCondition(conditions []infrastructurev1alpha3.CommonClusterStatusCondition) string {
 	if len(conditions) < 1 {
 		return naValue
 	}