+++ cmd/get/nodepools/runner_test.go
@@ -33,9 +33,9 @@ func Test_run(t *testing.T) {
 		{
 			name: "case 0: get nodepools",
 			storage: []runtime.Object{
-				newCAPIv1alpha2MachineDeployment("1sad2", "s921a", "2021-01-02T15:04:32Z", "10.5.0", 2, 1),
+				newCAPIv1alpha3MachineDeployment("1sad2", "s921a", "2021-01-02T15:04:32Z", "10.5.0", 2, 1),
 				newAWSMachineDeployment("1sad2", "s921a", "2021-01-01T15:04:32Z", "10.5.0", "test nodepool 3", 1, 3),
-				newCAPIv1alpha2MachineDeployment("f930q", "s921a", "2021-01-02T15:04:32Z", "11.0.0", 6, 6),
+				newCAPIv1alpha3MachineDeployment("f930q", "s921a", "2021-01-02T15:04:32Z", "11.0.0", 6, 6),
 				newAWSMachineDeployment("f930q", "s921a", "2021-01-02T15:04:32Z", "11.0.0", "test nodepool 4", 5, 8),
 			},
 			args:               nil,
@@ -50,9 +50,9 @@ func Test_run(t *testing.T) {
 		{
 			name: "case 2: get nodepool by name",
 			storage: []runtime.Object{
-				newCAPIv1alpha2MachineDeployment("1sad2", "s921a", "2021-01-02T15:04:32Z", "10.5.0", 2, 1),
+				newCAPIv1alpha3MachineDeployment("1sad2", "s921a", "2021-01-02T15:04:32Z", "10.5.0", 2, 1),
 				newAWSMachineDeployment("1sad2", "s921a", "2021-01-01T15:04:32Z", "10.5.0", "test nodepool 3", 1, 3),
-				newCAPIv1alpha2MachineDeployment("f930q", "s921a", "2021-01-02T15:04:32Z", "11.0.0", 6, 6),
+				newCAPIv1alpha3MachineDeployment("f930q", "s921a", "2021-01-02T15:04:32Z", "11.0.0", 6, 6),
 				newAWSMachineDeployment("f930q", "s921a", "2021-01-02T15:04:32Z", "11.0.0", "test nodepool 4", 5, 8),
 			},
 			args:               []string{"f930q"},
@@ -67,9 +67,9 @@ func Test_run(t *testing.T) {
 		{
 			name: "case 4: get nodepool by name, with no infrastructure ref",
 			storage: []runtime.Object{
-				newCAPIv1alpha2MachineDeployment("1sad2", "s921a", "2021-01-02T15:04:32Z", "10.5.0", 2, 1),
+				newCAPIv1alpha3MachineDeployment("1sad2", "s921a", "2021-01-02T15:04:32Z", "10.5.0", 2, 1),
 				newAWSMachineDeployment("1sad2", "s921a", "2021-01-01T15:04:32Z", "10.5.0", "test nodepool 3", 1, 3),
-				newCAPIv1alpha2MachineDeployment("f930q", "s921a", "2021-01-02T15:04:32Z", "11.0.0", 6, 6),
+				newCAPIv1alpha3MachineDeployment("f930q", "s921a", "2021-01-02T15:04:32Z", "11.0.0", 6, 6),
 			},
 			args:         []string{"f930q"},
 			errorMatcher: IsNotFound,
@@ -77,11 +77,11 @@ func Test_run(t *testing.T) {
 		{
 			name: "case 5: get nodepools by cluster name",
 			storage: []runtime.Object{
-				newCAPIv1alpha2MachineDeployment("1sad2", "s921a", "2021-01-02T15:04:32Z", "10.5.0", 2, 1),
+				newCAPIv1alpha3MachineDeployment("1sad2", "s921a", "2021-01-02T15:04:32Z", "10.5.0", 2, 1),
 				newAWSMachineDeployment("1sad2", "s921a", "2021-01-01T15:04:32Z", "10.5.0", "test nodepool 3", 1, 3),
-				newCAPIv1alpha2MachineDeployment("f930q", "s921a", "2021-01-02T15:04:32Z", "11.0.0", 6, 6),
+				newCAPIv1alpha3MachineDeployment("f930q", "s921a", "2021-01-02T15:04:32Z", "11.0.0", 6, 6),
 				newAWSMachineDeployment("f930q", "s921a", "2021-01-02T15:04:32Z", "11.0.0", "test nodepool 4", 5, 8),
-				newCAPIv1alpha2MachineDeployment("9f012", "29sa0", "2021-01-02T15:04:32Z", "9.0.0", 0, 3),
+				newCAPIv1alpha3MachineDeployment("9f012", "29sa0", "2021-01-02T15:04:32Z", "9.0.0", 0, 3),
 				newAWSMachineDeployment("9f012", "29sa0", "2021-01-02T15:04:32Z", "9.0.0", "test nodepool 5", 1, 1),
 			},
 			args:               nil,
@@ -91,11 +91,11 @@ func Test_run(t *testing.T) {
 		{
 			name: "case 6: get nodepools by name and cluster name",
 			storage: []runtime.Object{
-				newCAPIv1alpha2MachineDeployment("1sad2", "s921a", "2021-01-02T15:04:32Z", "10.5.0", 2, 1),
+				newCAPIv1alpha3MachineDeployment("1sad2", "s921a", "2021-01-02T15:04:32Z", "10.5.0", 2, 1),
 				newAWSMachineDeployment("1sad2", "s921a", "2021-01-01T15:04:32Z", "10.5.0", "test nodepool 3", 1, 3),
-				newCAPIv1alpha2MachineDeployment("f930q", "s921a", "2021-01-02T15:04:32Z", "11.0.0", 6, 6),
+				newCAPIv1alpha3MachineDeployment("f930q", "s921a", "2021-01-02T15:04:32Z", "11.0.0", 6, 6),
 				newAWSMachineDeployment("f930q", "s921a", "2021-01-02T15:04:32Z", "11.0.0", "test nodepool 4", 5, 8),
-				newCAPIv1alpha2MachineDeployment("9f012", "29sa0", "2021-01-02T15:04:32Z", "9.0.0", 0, 3),
+				newCAPIv1alpha3MachineDeployment("9f012", "29sa0", "2021-01-02T15:04:32Z", "9.0.0", 0, 3),
 				newAWSMachineDeployment("9f012", "29sa0", "2021-01-02T15:04:32Z", "9.0.0", "test nodepool 5", 1, 1),
 			},
 			args:               []string{"f930q"},