+++ cmd/template/networkpool/provider/common.go
@@ -4,7 +4,7 @@ import (
 	"io"
 	"text/template"
 
-	"github.com/giantswarm/apiextensions/v3/pkg/apis/infrastructure/v1alpha2"
+	"github.com/giantswarm/apiextensions/v3/pkg/apis/infrastructure/v1alpha3"
 
 	"github.com/giantswarm/microerror"
 	"sigs.k8s.io/yaml"
@@ -22,13 +22,13 @@ type NetworkPoolCRsConfig struct {
 func WriteTemplate(out io.Writer, config NetworkPoolCRsConfig) error {
 	var err error
 
-	crsConfig := v1alpha2.NetworkPoolCRsConfig{
+	crsConfig := v1alpha3.NetworkPoolCRsConfig{
 		CIDRBlock:     config.CIDRBlock,
 		NetworkPoolID: config.NetworkPoolName,
 		Owner:         config.Owner,
 	}
 
-	crs, err := v1alpha2.NewNetworkPoolCRs(crsConfig)
+	crs, err := v1alpha3.NewNetworkPoolCRs(crsConfig)
 	if err != nil {
 		return microerror.Mask(err)
 	}