+++ pkg/data/domain/nodepool/aws.go
@@ -3,11 +3,11 @@ package nodepool
 import (
 	"context"
 
-	infrastructurev1alpha2 "github.com/giantswarm/apiextensions/v3/pkg/apis/infrastructure/v1alpha2"
+	infrastructurev1alpha3 "github.com/giantswarm/apiextensions/v3/pkg/apis/infrastructure/v1alpha3"
 	"github.com/giantswarm/apiextensions/v3/pkg/label"
 	"github.com/giantswarm/microerror"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	capiv1alpha2 "sigs.k8s.io/cluster-api/api/v1alpha2"
+	capiv1alpha3 "sigs.k8s.io/cluster-api/api/v1alpha3"
 	runtimeClient "sigs.k8s.io/controller-runtime/pkg/client"
 )
 
@@ -20,9 +20,9 @@ func (s *Service) getAllAWS(ctx context.Context, namespace, clusterID string) (R
 	}
 	inNamespace := runtimeClient.InNamespace(namespace)
 
-	var awsMDs map[string]*infrastructurev1alpha2.AWSMachineDeployment
+	var awsMDs map[string]*infrastructurev1alpha3.AWSMachineDeployment
 	{
-		mdCollection := &infrastructurev1alpha2.AWSMachineDeploymentList{}
+		mdCollection := &infrastructurev1alpha3.AWSMachineDeploymentList{}
 		err = s.client.K8sClient.CtrlClient().List(ctx, mdCollection, labelSelector, inNamespace)
 		if err != nil {
 			return nil, microerror.Mask(err)
@@ -30,14 +30,14 @@ func (s *Service) getAllAWS(ctx context.Context, namespace, clusterID string) (R
 			return nil, microerror.Mask(noResourcesError)
 		}
 
-		awsMDs = make(map[string]*infrastructurev1alpha2.AWSMachineDeployment, len(mdCollection.Items))
+		awsMDs = make(map[string]*infrastructurev1alpha3.AWSMachineDeployment, len(mdCollection.Items))
 		for _, machineDeployment := range mdCollection.Items {
 			md := machineDeployment
 			awsMDs[machineDeployment.GetName()] = &md
 		}
 	}
 
-	machineDeployments := &capiv1alpha2.MachineDeploymentList{}
+	machineDeployments := &capiv1alpha3.MachineDeploymentList{}
 	{
 		err = s.client.K8sClient.CtrlClient().List(ctx, machineDeployments, labelSelector, inNamespace)
 		if err != nil {
@@ -54,10 +54,10 @@ func (s *Service) getAllAWS(ctx context.Context, namespace, clusterID string) (R
 
 			if awsMD, exists := awsMDs[cr.GetName()]; exists {
 				cr.TypeMeta = metav1.TypeMeta{
-					APIVersion: "cluster.x-k8s.io/v1alpha2",
+					APIVersion: "cluster.x-k8s.io/v1alpha3",
 					Kind:       "MachineDeployment",
 				}
-				awsMD.TypeMeta = infrastructurev1alpha2.NewAWSMachineDeploymentTypeMeta()
+				awsMD.TypeMeta = infrastructurev1alpha3.NewAWSMachineDeploymentTypeMeta()
 
 				np := Nodepool{
 					MachineDeployment:    &o,
@@ -85,7 +85,7 @@ func (s *Service) getByIdAWS(ctx context.Context, id, namespace, clusterID strin
 	np := &Nodepool{}
 
 	{
-		crs := &capiv1alpha2.MachineDeploymentList{}
+		crs := &capiv1alpha3.MachineDeploymentList{}
 		err = s.client.K8sClient.CtrlClient().List(ctx, crs, labelSelector, inNamespace)
 		if err != nil {
 			return nil, microerror.Mask(err)
@@ -97,13 +97,13 @@ func (s *Service) getByIdAWS(ctx context.Context, id, namespace, clusterID strin
 		np.MachineDeployment = &crs.Items[0]
 
 		np.MachineDeployment.TypeMeta = metav1.TypeMeta{
-			APIVersion: "cluster.x-k8s.io/v1alpha2",
+			APIVersion: "cluster.x-k8s.io/v1alpha3",
 			Kind:       "MachineDeployment",
 		}
 	}
 
 	{
-		crs := &infrastructurev1alpha2.AWSMachineDeploymentList{}
+		crs := &infrastructurev1alpha3.AWSMachineDeploymentList{}
 		err = s.client.K8sClient.CtrlClient().List(ctx, crs, labelSelector, inNamespace)
 		if err != nil {
 			return nil, microerror.Mask(err)
@@ -114,7 +114,7 @@ func (s *Service) getByIdAWS(ctx context.Context, id, namespace, clusterID strin
 		}
 		np.AWSMachineDeployment = &crs.Items[0]
 
-		np.AWSMachineDeployment.TypeMeta = infrastructurev1alpha2.NewAWSMachineDeploymentTypeMeta()
+		np.AWSMachineDeployment.TypeMeta = infrastructurev1alpha3.NewAWSMachineDeploymentTypeMeta()
 	}
 
 	return np, nil